LazyReference fix for not setting it to resolve to soon if the type is a IRIValueType because the function collection doesn't have to be there yet.
drop the creation of Type objects in LazyTypeReference, now always just
create IRIValueType for custom/local resolved types.

handleDeclarations can't do RModelBuilder.create(getContext(), method)
or variable just yet when initalizing just the Function or Variable.
first all Functions or Variables must be there to be able to resolve all
the types

Let SimpleType first try to resolve it and if it is then an UNKNOWN type
then only try to create a IRIValueType. Also if because of some
constructs a still a simple type is created that should have been a
IRIValueType then just test on name to be assignable

added test for the lazy stuff.