For SVY-17921: Adding missing String methods that are supported by Rhino
Made JSLib (so native javascript types) show params/sample/return value
besides description in Solution Explorer view + added code that can
generate a description for such a method in order to copy/past it to the
dltk native-references.xml for code completion. That file will is
updated with the newly added/changed docs from this case; but as that is
currently a manual process, a  followup will be created to update the
code completion method descriptions for all existing jsdocs that are
better in our doc java classes.