ValueCollectionFactor.createValueCollection now returns a IValueCollectionReference
which is an indirection because the collection can change to immutable a
bit later.

The Factory now holds in own stack for the current thread and only will
make them all Immutable when the stack unwinds. So all Collections that
are depending on each other will be made immutable in one go.
This way types accross files that have cycles in them will be able to
resolve each other because the make immutable will be done when all of
them have still full pointers to each other.