SVY-20596 add type support in (arrow) functions that are used in specific "foreach" functions of Array/Map/Set
introduced a lot of support for the forEach() and then typing the
function parameters that are given including also showing the
description of what the callback should be.

made Map and Set keys() and values() function generic returning now the
Generic Iterable that by itself returns a generic IterableValue

fixed various problems i the selection engine so it does show in the
tooltip the correct type.

RMapType is now really mapping on the Map type so it will show
Map<String,String> or Set<String> instead of Object<String>

to show the correct generic type the IRMethod and IRProperty values are
now passed through, instead of the IModelElement variants because those
lose the actual runtime type.
This does result that code completion will now give a much descriptive
completion (so including the full types)