SVY-20548 Add support for JSComponent.<componentName> and componentProperties / handlers
introduced a MapOnType that IRType can implement
So an RType can say to the system that it maps at runtime on a String
also, but when code completion this is not know and should not be
visible. But the validation does work because it will check if the
runtime map type is compartiable.